Conflicts:
	atm program/atm program.sdf
	test - project/Debug/test project.dll
	test - project/Debug/test project.pdb
	test - project/test project.sdf
	test - project/test project.v12.suo
	test - project/test project/Debug/Account.obj
	test - project/test project/Debug/AssemblyInfo.obj
	test - project/test project/Debug/Stdafx.obj
	test - project/test project/Debug/User.obj
	test - project/test project/Debug/test project.log
	test - project/test project/Debug/test project.obj
	test - project/test project/Debug/test project.pch
	test - project/test project/Debug/test project.tlog/CL.read.1.tlog
	test - project/test project/Debug/test project.tlog/CL.write.1.tlog
	test - project/test project/Debug/test project.tlog/cl.command.1.tlog
	test - project/test project/Debug/test project.tlog/link.command.1.tlog
	test - project/test project/Debug/test project.tlog/link.read.1.tlog
	test - project/test project/Debug/test project.tlog/link.write.1.tlog
	test - project/test project/Debug/test project.tlog/metagen.read.1.tlog
	test - project/test project/Debug/test project.tlog/metagen.write.1.tlog
	test - project/test project/Debug/test project.tlog/rc.command.1.tlog
	test - project/test project/Debug/test project.tlog/rc.read.1.tlog
	test - project/test project/Debug/test project.tlog/rc.write.1.tlog
	test - project/test project/Debug/test project.tlog/test project.write.1u.tlog
	test - project/test project/Debug/vc120.pdb

+++ test - project/test project/User.cpp
@@ -44,7 +44,7 @@ bool User::setDob(std::string DOB)
 
 bool User::addCard(int Card)
 {
-	int cardNum[3] = {123456789123, 987654321987, 654321987654};
+	long int cardNum[3] = {123456789123, 987654321987, 654321987654};
 
 	for (int i = 0; i < 3; i++)
 	{
